Edit .gitignore Library entry
Changed /Library/ to Library/ in order to allow git to see the Library
folder in the Unity Dungeon subdirectory. This change was needed as
the Unity project is in its own folder instead of in the main repo
directory causing Library files to still be tracked as it was previously
only checking the main repo director.