Add new camera functionalities and update world
I added a script that makes the camera follow the player and
restricts it from leaving the map. I also added an effect that
makes it look as if the camera is trying to catch up with the
player because I thought it was cool. Then I just added some
more assets to the world to make it look cooler.