Add new room with camera transition and Prefabs
I added a second tilemap with the same features as the first
(Room1 and Room2). I then made a script to transition the
player and the camera into the second room when a trigger
is hit. I then figured out prefabs and made a mutable
prefab folder and added the trigger into it. There are
tons of things we need to go back to and put them into
the prefabs because I just learned how it works and why
it is helpful.