Add text UI with title cards for all rooms
I added an empty UI game object for text, centered it,
made it inactive and added code that allows you to
title each room. The code also activated the text once
the room is entered and then it goes away after a period
of time. I added all of that code in the same cam switch
code as the previous commit.