Add IceMage movement animations to Player
Added various IceMage animation clips and the IceMage animation
controller. Controller determines current horizonal input and saves it
into the lastDirection variable. This variable is only updated when the
player is moving horizontally in some way. This is used to keep facing
previous direction when idling or only moving vertically.