Add example scene with branching text
Add ink add on with ink story, added an example
scene for later reference on how to make text
with branching options. A lot of work went into
trying to copy the tutorial I was following
but i still dont really understand a lot of it.
It is mainly for us to reference later.