Add example platforming mini level
I added a new scene with a new prefab Player
with a playercontroler script that I got from
git hub. I also made a new movement script to allow for
platforming mechanics. I then added a new room in the main scene
with a trigger that brings you to the platforming level

I did a lot of this with my own creativity bro suck me off

I only used a tutorial for the movement mechanics bitch